+++ src/mw-ui-components/directives/mw_tab_bar.js
@@ -1,61 +1,128 @@
 angular.module('mwUI.UiComponents')
-  //TODO rename
-  /**
-   * @example ```html
-   * <!-- change callback example -->
-   * <div mw-tabs active-pane-number="myCtrl.activePane" tab-changed="myCtrl.tabChanged">
-      <div mw-tabs-pane="{{'mytitle'| i18n}}">
-      Tab 1
-      </div>
-      <div mw-tabs-pane="{{'mytitle_2'| i18n}}">
-      Tab 2
-      </div>
-    </div>
-   * ```
-   */
-  .directive('mwTabs', function () {
+//TODO rename
+/**
+ * @example ```html
+ * <!-- change callback example -->
+ * <div mw-tabs active-pane-number="myCtrl.activePane" tab-changed="myCtrl.tabChanged">
+ <div mw-tabs-pane="{{'mytitle'| i18n}}">
+ Tab 1
+ </div>
+ <div mw-tabs-pane="{{'mytitle_2'| i18n}}">
+ Tab 2
+ </div>
+ </div>
+ * ```
+ */
+  .directive('mwTabs', function ($rootScope) {
     return {
       transclude: true,
       scope: {
-        justified: '=',
-        activePaneNumber: '=',
-        tabChanged: '='
+        justified: '=?',
+        activePaneNumber: '=?',
+        activePaneId: '@?',
+        tabChanged: '=?'
       },
       templateUrl: 'uikit/mw-ui-components/directives/templates/mw_tab_bar.html',
       controller: function ($scope) {
-        var panes = $scope.panes = [];
-        $scope.select = function (pane) {
-          angular.forEach(panes, function (p) {
-            p.selected = false;
-          });
+        // The panes are populated by the mw-tabs-pane directive that is calling `registerPane`of this controller
+        // It is defined as a scope attribute so it is available in this html template
+        var panes = $scope.panes = [],
+          activePaneIndex;
 
-          if($scope.activePaneNumber){
-            $scope.activePaneNumber = _.indexOf($scope.panes,pane)+1;
-          }
+        var setInitialSelection = function () {
+          activePaneIndex = null;
 
-          pane.selected = true;
-          // emit the callback
-          if ($scope.tabChanged && typeof $scope.tabChanged === 'function') {
-            $scope.tabChanged($scope.activePaneNumber);
+          // In case that no active pane is defined by setting activePaneNumber or Id the first pane will be selected
+          if (angular.isUndefined($scope.activePaneNumber) && angular.isUndefined($scope.activePaneId) && panes.length>0) {
+            $scope.select(panes[0]);
+
+          // When a pane number is defined the pane with the index of activePaneNumber + 1 will be selected
+          // So when the second tab shall be selected set activePaneNumber to 2
+          } else if (angular.isDefined($scope.activePaneNumber)) {
+            $scope.selectTabByNumber($scope.activePaneNumber);
+
+          // When a pane id is defined the pane where the pane id equals the activePaneId will be selected
+          // Make sure to set an id on the mw-tab-pane
+          } else if (angular.isDefined($scope.activePaneId)) {
+            $scope.selectTabById($scope.activePaneId);
           }
         };
-        
-        // add a change listener on the pane 
-        if ($scope.tabChanged && typeof $scope.tabChanged === 'function') { 
-          $scope.$watch('activePaneNumber', function (_new, _old) {
-            if (_new !== _old) {
-              $scope.select(panes[_new - 1]);
+
+        var throttledSetInitialSelection = _.debounce(setInitialSelection, 100);
+
+        $scope.getActivePane = function () {
+          var activePane;
+          $scope.panes.every(function (pane) {
+            if (pane.isSelected()) {
+              activePane = pane;
+              return false;
             }
+            return true;
           });
-        }
+          return activePane;
+        };
+
+        $scope.select = function (newPane) {
+          var newPaneIndex = _.indexOf($scope.panes, newPane);
+
+          if (newPane && newPaneIndex !== -1 && newPaneIndex !== activePaneIndex) {
+            var previousSelectedPane = $scope.getActivePane();
+            if (previousSelectedPane) {
+              previousSelectedPane.deselect();
+            }
+
+            newPane.select();
+            activePaneIndex = newPaneIndex;
+
+            if ($scope.tabChanged && typeof $scope.tabChanged === 'function') {
+              $scope.tabChanged(activePaneIndex + 1, newPane, previousSelectedPane);
+              $rootScope.$emit('$mwTabChange');
+            }
+          }
+        };
+
+        $scope.selectTabByNumber = function (number) {
+          if (number > 0 && number <= $scope.tabs.length) {
+            $scope.select(panes[number - 1]);
+          }
+        };
+
+        $scope.selectTabById = function (id) {
+          $scope.panes.every(function (pane) {
+            if (pane.getId() === id) {
+              $scope.select(pane);
+              return false;
+            }
+            return true;
+          });
+        };
+
+        // add a change listener on the pane
+        $scope.$watch('activePaneNumber', function (_new, _old) {
+          if (_new && _new !== _old) {
+            $scope.selectTabByNumber(_new);
+          }
+        });
+
+        $scope.$watch('activePaneId', function (_newId) {
+          if (_newId) {
+            $scope.selectTabById(_newId);
+          }
+        });
 
         this.registerPane = function (pane) {
-          if ( ( $scope.activePaneNumber && $scope.activePaneNumber-1 === panes.length) || (!panes.length && !$scope.activePaneNumber) ) {
-            var bak = $scope.activePaneNumber;
-            $scope.select(pane);
-            $scope.activePaneNumber = bak;
-          }
           panes.push(pane);
+          throttledSetInitialSelection();
+        };
+
+        this.unRegisterPane = function (pane) {
+          if (pane) {
+            var indexOfExistingPane = _.indexOf(panes, pane);
+            if (indexOfExistingPane !== -1) {
+              $scope.panes.splice(indexOfExistingPane, 1);
+            }
+            throttledSetInitialSelection();
+          }
         };
       }
     };